Implement course detail

+++ lib/common/widget/base_text_widget.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:ulearning_app/common/values/colors.dart';
 
-Widget reusableText(String text, {Color color = AppColors.primaryText,int fontSize = 16,FontWeight fontWeight = FontWeight.bold}){
+Widget reusableText(String text, {Color color = AppColors.primaryText,double fontSize = 16,FontWeight fontWeight = FontWeight.bold}){
   return  Text(
     text,
     style: TextStyle(
